fix(clock,intervals): fix clock race condition
if on the recumples program, you pressed, after the full reset, step
three times (the first instruction has only 4 cycles), and then step
instruction, the program would go on until the end of the program.
This is because the first invocation to the timer function
(sendCycleEvent) occurs within setTimer itself, which means it hasn't
returned yet, and thus hasn't set this.interval_id, and during that
cycle, the instruction finishes, and the clock module tries to clear a
timer. Because the this.interval_id is not yet set, it fails (it doesn't
know which timer it has).